libs/python/voxel-classic/pyproject.toml
[project]
name = "voxel-classic"
readme = "README.md"
version = "0.1.0"
description = "A python package for generalized microscope control"
license = { file = "LICENSE" }
# requires-python = ">=3.10, <=3.12"
requires-python = ">=3.12, <3.13"
authors = [
    { name = "Micah Woodard", email = "micah.woodard@alleninstitute.org" },
    { name = "Adam Glaser", email = "adam.glaser@alleninstitute.org" },
    { name = "Walter Mwaniki", email = "walter.mwaniki@alleninstitute.org" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

keywords = ["microscopy", "microscope", "acquisition", "imaging"]

dependencies = [
    "ruamel-yaml>=0.18.5",
    "inflection>=0.5.1",
    "pydantic>=2.7.3",
    "imohash>=1.0.5",
    "numpy>=2.1",
    # device driver dependencies
    "pyserial>=3.5",
    "pyusb>=1.2.1",
    "psutil>=5.9.8",
    "PyVISA>=1.14.1",
    "PyVISA-py>=0.7.2",
    "pylablib>=1.4.2",
    # "nidaqmx==1.0.0",
    "nidaqmx>=1.0.1",
    # data handling dependencies
    "h5py>=3.0.0",
    # processing and utils dependencies
    "pyclesperanto>=0.8.2",
    "scikit-image>=0.20.0",
    "gputools>0.2.14",
    "tensorstore>=0.1.56",
    "matplotlib>=3.8.2",
    "scipy>=1.12.0",
    "fast_histogram>=0.14",
    "sympy>=1.12.1",
    "numpy>2.0.0",
    "siphash24>=1.7",
]


[project.optional-dependencies]
bdv = ["h5py>=3.12.1", "tqdm>=4.66.5"]
imaris = ["pyimariswriter>=0.7.0"]
tifffile-writer = ["tifffile>=2024.1.30"]
acquire-zarr-writer = ["acquire-zarr>=0.0.5"]
writers = ["voxel-classic[bdv, imaris, tifffile-writer, acquire-zarr-writer]"]

ximea-camera = [
    "ximea-python @ https://github.com/adamkglaser/ximea-python/releases/download/v4.29.10/ximea_python-4.29.10-py3-none-any.whl",
]
cameras = ["voxel-classic[ximea-camera]"]
tigerasi-controller = ["tigerasi"]
scikit-tensor = [
    "scikit-tensor-py3 @ https://github.com/adamkglaser/scikit-tensor-py3/releases/download/v0.1.0/scikit_tensor_py3-0.4.1-py3-none-any.whl",
]

cobolt = ["pycobolt"]
coherent = ["coherent-lasers"]
vortran = ["vortran-laser"]
oxxius = ["oxxius-laser"]
obis = ["obis-laser"]
from-git = ["voxel-classic[cobolt, coherent, vortran, oxxius, obis]"]

aaopto = ["aaopto-aotf>=0.0.11"]
etl-opto = ["optoicc", "optokummenberg"]
egrabber = ["egrabber"]
from-wheelhouse = ["voxel-classic[aaopto, etl-opto, egrabber]"]

# [tool.uv.sources]
# pycobolt = { git = "https://github.com/cobolt-lasers/pycobolt.git" }
# obis-laser = { git = "https://github.com/AllenNeuralDynamics/obis-laser.git" }
# coherent-lasers = { git = "https://github.com/AllenNeuralDynamics/coherent-lasers.git" }
# oxxius-laser = { git = "https://github.com/AllenNeuralDynamics/oxxius_laser.git" }
# vortran-laser = { git = "https://github.com/AllenNeuralDynamics/vortran_laser.git" }
# optokummenberg = { path = ".wheels/optoKummenberg-1.0.4894-py3-none-any.whl" }
# optoicc = { path = ".wheels/optoICC-2.0.4922-py3-none-any.whl" }
# egrabber = { path = ".wheels/egrabber-24.09.0.5-py2.py3-none-any.whl" }


[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
